+++ modules/@apostrophecms/area/ui/apos/components/AposAreaWidget.vue
@@ -112,6 +112,8 @@
           :disabled="disabled"
           :max-reached="maxReached"
           :tabbable="isFocused"
+          :model-value="widget"
+          :area-field="field"
           @up="$emit('up', i);"
           @remove="$emit('remove', i);"
           @edit="$emit('edit', i);"