+++ modules/@apostrophecms/modal/ui/apos/components/AposWidgetModalTabs.vue
@@ -139,7 +139,7 @@ export default {
       this.$emit('select-tab', id);
     },
     moreMenuHandler(item) {
-      this.$emit('select-tab', item);
+      this.$emit('select-tab', item.action);
     }
   }
 };